ml-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
opencv-python-headless==4.9.0.80
Pillow==10.3.0
imagehash==4.3.1
scikit-surprise==1.1.4
nltk==3.8.1
crewai==0.30.11
langchain==0.2.1
langchain-anthropic==0.1.13
python-dotenv==1.0.1
requests==2.32.3
